feat: add rust treesitter indexing
