+++ internal/indexing/treesitter_stub.go
@@ -1,4 +1,5 @@
 //go:build !cgo
+// +build !cgo
 
 package indexing
